+++ pic24f-lcd-curiosity-oob.X/main.c
@@ -13,11 +13,11 @@
   @Description
     This source file provides main entry point for system initialization and application code development.
     Generation Information :
-        Product Revision  :  PIC24 / dsPIC33 / PIC32MM MCUs - 1.165.1
+        Product Revision  :  PIC24 / dsPIC33 / PIC32MM MCUs - 1.171.0
         Device            :  PIC24FJ128GL306
     The generated drivers are tested against the following:
-        Compiler          :  XC16 v1.41
-        MPLAB 	          :  MPLAB X v5.30
+        Compiler          :  XC16 v1.70
+        MPLAB 	          :  MPLAB X v5.50
 */
 
 /*
@@ -45,99 +45,24 @@
 /**
   Section: Included Files
 */
-
-#include <stddef.h>
-
-#include "bsp/power.h"
-#include "application/usb_operational_mode.h"
-#include "application/battery_operational_mode.h"
-#include "mcc_generated_files/rtcc.h"
-#include "bsp/power.h"
-#include "application/lcd_demo.h"
-#include "bsp/build_time.h"
 #include "mcc_generated_files/system.h"
-#include "mcc_generated_files/lcd_segments.h"
 
-static void SwitchOperatoinalMode(enum POWER_SOURCE new_source);
-
-static enum POWER_SOURCE current_source;
-static const struct OPERATIONAL_MODE *operational_mode;
-
-/*******************************************************************************
-  GETTING STARTED
-  -----------------------------------------------------------------------------
-  To run this demo, please refer to the readme.txt file that is provided with
-  this project.  You can find this file attached to the project in the 
-  "Documentation" logical folder in the project view in the IDE.
-  
-  You can also locate the readme.txt file on next to the project folder where
-  this demo was extracted.  
-  
-  The readme.txt contains the details of how to run the demo.
-  
-  There is an additional hardware.txt file also in the "Documentation" logical
-  folder in the project that summarizes the hardware connections of the board.
-  For a more detailed hardware description, please refer to the boards user's 
-  guide for a full schematic at the board website:
-  www.microchip.com/pic24flcdcuriosity 
- ******************************************************************************/
-
+/*
+                         Main application
+ */
 int main(void)
 {
-    struct tm build_time;
-    
-    enum POWER_SOURCE new_source;
-    
+    // initialize the device
     SYSTEM_Initialize();
-    
-    LCD_MICROCHIP1_On();
-    BUILDTIME_Get(&build_time);
-    RTCC_TimeSet(&build_time);
-    
-    operational_mode = NULL;
-    current_source = POWER_SOURCE_UNKNOWN;
-    new_source = POWER_GetSource();
-    
-    while(1)
-    {
-        if(new_source != current_source)
-        {
-            current_source = new_source;
-            SwitchOperatoinalMode(new_source);
-        }
-        do
-        {
-            operational_mode->Tasks();
-            new_source = POWER_GetSource();
-            
-        } while( current_source == new_source );
-    }
-    
-    return 0;
-}
 
-static void SwitchOperatoinalMode(enum POWER_SOURCE new_source)
-{
-    if(operational_mode != NULL)
+    while (1)
     {
-        operational_mode->Deinitialize();
+        // Add your application code
     }
-    
-    switch(new_source)
-    {
-        case POWER_SOURCE_USB:
-            operational_mode = &usb_operational_mode;
-            break;
 
-        case POWER_SOURCE_BATTERY:
-            operational_mode = &battery_operational_mode;
-            break;
-            
-        default:
-            break;
-    }
-    if(operational_mode != NULL)
-    {
-        operational_mode->Initialize();
-    }
+    return 1;
 }
+/**
+ End of File
+*/
+