+++ pic24f-lcd-curiosity-oob.X/mcc_generated_files/mcc.h
@@ -13,11 +13,11 @@
   @Description:
     This file will be removed in future MCC releases. Use system.h instead.
     Generation Information :
-        Product Revision  :  PIC24 / dsPIC33 / PIC32MM MCUs - 1.169.0
+        Product Revision  :  PIC24 / dsPIC33 / PIC32MM MCUs - 1.171.0
         Device            :  PIC24FJ128GL306
     The generated drivers are tested against the following:
-        Compiler          :  XC16 v1.50
-        MPLAB             :  MPLAB X v5.40
+        Compiler          :  XC16 v1.70
+        MPLAB             :  MPLAB X v5.50
 */
 
 /*
@@ -51,17 +51,17 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-#include "lcd.h"
 #include "adc1.h"
-#include "spi1.h"
 #include "rtcc.h"
+#include "uart1.h"
+#include "interrupt_manager.h"
+#include "traps.h"
+#include "lcd.h"
 #include "tmr3.h"
 #include "mccp2_compare.h"
 #include "mccp3_compare.h"
 #include "mccp4_compare.h"
-#include "interrupt_manager.h"
-#include "traps.h"
-#include "uart1.h"
+#include "spi1.h"
 
 #warning "This file will be removed in future MCC releases. Use system.h instead."