+++ changelog.md
@@ -1,3 +1,8 @@
+# pic24f-lcd-curiosity-oob v1.0.2
+### Release Highlights
+
+Patch release to fix the build error issue with the latest tool pack.
+
 # pic24f-lcd-curiosity-oob v1.0.1
 ### Release Highlights